Support iluvatar pretrain.

+++ paddlenlp/transformers/llama/modeling.py
@@ -1260,7 +1260,8 @@ class LlamaModel(LlamaPretrainedModel):
         else:
             expanded_attn_mask = _make_causal_mask(input_shape, past_key_values_length=past_key_values_length)
         # Convert bool attention_mask to float attention mask, which will be added to attention_scores later
-        expanded_attn_mask = paddle.where(expanded_attn_mask, 0.0, paddle.finfo(dtype).min).astype(dtype)
+        import numpy as np
+        expanded_attn_mask = paddle.where(expanded_attn_mask, np.float32(0.0), np.float32(paddle.finfo(dtype).min)).astype(dtype)
         return expanded_attn_mask
 
     @paddle.jit.not_to_static
